lstm: Fix some signed/unsigned compiler warnings

+++ src/lstm/plumbing.cpp
@@ -142,7 +142,7 @@ void Plumbing::DebugWeights() {
 
 // Returns a set of strings representing the layer-ids of all layers below.
 void Plumbing::EnumerateLayers(const std::string *prefix, std::vector<std::string> &layers) const {
-  for (int i = 0; i < stack_.size(); ++i) {
+  for (size_t i = 0; i < stack_.size(); ++i) {
     std::string layer_name;
     if (prefix) {
       layer_name = *prefix;
@@ -238,7 +238,7 @@ bool Plumbing::DeSerialize(TFile *fp) {
 // Updates the weights using the given learning rate, momentum and adam_beta.
 // num_samples is used in the adam computation iff use_adam_ is true.
 void Plumbing::Update(float learning_rate, float momentum, float adam_beta, int num_samples) {
-  for (int i = 0; i < stack_.size(); ++i) {
+  for (size_t i = 0; i < stack_.size(); ++i) {
     if (network_flags_ & NF_LAYER_SPECIFIC_LR) {
       if (i < learning_rates_.size()) {
         learning_rate = learning_rates_[i];
@@ -259,7 +259,7 @@ void Plumbing::CountAlternators(const Network &other, TFloat *same, TFloat *chan
   ASSERT_HOST(other.type() == type_);
   const auto *plumbing = static_cast<const Plumbing *>(&other);
   ASSERT_HOST(plumbing->stack_.size() == stack_.size());
-  for (int i = 0; i < stack_.size(); ++i) {
+  for (size_t i = 0; i < stack_.size(); ++i) {
     stack_[i]->CountAlternators(*plumbing->stack_[i], same, changed);
   }
 }
